Sync demos in docs + change location in a separated folder
