--- components/content/inspira/examples/backgrounds/DotBackgroundDemo.vue
@@ -1,9 +1,0 @@
-<template>
-  <DotBackground :animate="true">
-    <p
-      className="text-4xl sm:text-5xl font-bold relative z-20 bg-clip-text text-transparent bg-gradient-to-b from-neutral-200 to-neutral-500 py-8"
-    >
-      Dot Background
-    </p>
-  </DotBackground>
-</template>
