--- components/content/inspira/examples/backgrounds/GridBackgroundDemo.vue
@@ -1,9 +1,0 @@
-<template>
-  <GridBackground :animate="true" :direction="'top-right'">
-    <p
-      className="text-4xl sm:text-5xl font-bold relative z-20 bg-clip-text text-transparent bg-gradient-to-b from-neutral-200 to-neutral-500 py-8"
-    >
-      Grid Background
-    </p>
-  </GridBackground>
-</template>
